Cleanup SkillManager
Fix IsLearnableSkill
