+++ src/CorePluginAPI/EPlayerClass.cs
@@ -1,0 +1,9 @@
+﻿namespace QuantumCore.API;
+
+public enum EPlayerClass
+{
+    Warrior = 0,
+    Ninja = 1,
+    Sura = 2,
+    Shaman = 3,
+}
